Use double instead of float, per the existing APIs.
Also improve documentation.

+++ include/client/mir_toolkit/events/surface_output_event.h
@@ -55,13 +55,15 @@ MirFormFactor mir_surface_output_event_get_form_factor(MirSurfaceOutputEvent con
 float mir_surface_output_event_get_scale(MirSurfaceOutputEvent const* ev);
 
 /**
- * Retrieve the refresh rate of the new output configuration of a
- * MirSurfaceOutputEvent.
+ * Retrieve the maximum refresh rate of the output(s) associated with a
+ * MirSurfaceOutputEvent. For variable refresh rate displays this represents
+ * the maximum refresh rate of the display to aim for, rather than a measurement
+ * of recent performance.
  *
  * \param [in] ev   The event
  * \return          The refresh rate in Hz
  */
-float mir_surface_output_event_get_refresh_rate(MirSurfaceOutputEvent const* ev);
+double mir_surface_output_event_get_refresh_rate(MirSurfaceOutputEvent const* ev);
 
 /**
  * Retrieve the ID of the output this surface is on from a MirSurfaceOutputEvent
