+++ src/server/scene/output_properties_cache.h
@@ -37,7 +37,7 @@ struct OutputProperties
     geometry::Rectangle extents;
     int dpi;
     float scale;
-    float refresh_rate;
+    double refresh_rate;
     MirFormFactor form_factor;
     graphics::DisplayConfigurationOutputId id;
 };
